third_party/blink/web_tests/external/wpt/css/css-position/sticky/position-sticky-horizontal-rtl-overconstrained-expected.txt

This is a testharness.js-based test.
[FAIL] start of scroll container
  assert_equals: expected 0 but got 200
[FAIL] right before the in-flow position of the sticky box
  assert_equals: expected 80 but got 200
[FAIL] right after the in-flow position the sticky box
  assert_equals: expected 120 but got 220
Harness: the test ran to completion.
